Added inital vehicle controls with a testing scene
The controls right now are very slippery and completely require a
controller to use effectively. There are bindings for keyboard+mouse,
but they're not very good at the moment. The script that's there serves
more as a base for testing and improvement than it does anything final.
Everything should currently natively work with an XBox 360 controller.